limit of algorithm = cpu time

+++ Reading_Data.py
@@ -11,6 +11,6 @@ def Reading(file_name):
     for i in range(nK):
         Weights.append([float(x) for x in data.readline().split(" ")])
     data.readlines(2)
-    real_answer = int(data.readline())    
+    real_answer = float(data.readline())    
         
     return nK, nI, Capacity, Profits, Weights, real_answer
